glove: move LVGL-related code into chart.cc
Now main.cc is only 25 lines of code and reads like a table of contents.
