+++ glove/main/CMakeLists.txt
@@ -1,1 +1,2 @@
-idf_component_register(SRCS main.cc button.cc buttons.cc INCLUDE_DIRS .)
+idf_component_register(SRCS main.cc button.cc buttons.cc chart.cc
+                       INCLUDE_DIRS .)
