+++ glove/main/chart.hh
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <lvgl.h>
+
+void lvgl_init();
+lv_obj_t *chart_init();
